@@ -4,7 +4,7 @@ publish_to: 'none'
 version: 2.0.16+0
 
 environment:
-  sdk: '>=3.4.1 <4.0.0'
+  sdk: '>=3.10.0 <4.0.0'
 dependencies:
   flutter:
     sdk: flutter
@@ -71,9 +71,7 @@ dependencies:
   bloc: ^9.2.0
   meta: ^1.17.0
   loading_indicator_m3e: ^0.1.1
-  m3e_design: ^0.2.1
-  expressive_refresh: ^0.1.2
-  navigation_rail_m3e: ^0.3.5
+  # m3e_collection: ^0.3.7
   package_info_plus: ^9.0.0
   flutter_markdown_plus: ^1.0.7
   yt_music:
@@ -81,6 +79,9 @@ dependencies:
       url: https://github.com/sheikhhaziq/yt_music.git
   connectivity_plus: ^7.0.0
   audio_service_mpris: ^0.2.0
+  fluentui_system_icons: ^1.1.273
+  navigation_rail_m3e: ^0.3.5
+  m3e_collection: ^0.3.7
 
 dev_dependencies:
   flutter_test:
