@@ -76,7 +76,7 @@ class AdaptiveListTile extends StatelessWidget {
                   const EdgeInsets.symmetric(horizontal: 16.0, vertical: 8),
               decoration: BoxDecoration(
                 color: selected
-                    ? theme.colorScheme.primary.withOpacity(0.1)
+                    ? theme.colorScheme.primary.withValues(alpha:0.1)
                     : backgroundColor,
               ),
               child: Column(
