@@ -1,18 +1,19 @@
 import 'dart:collection';
 
 import 'package:cached_network_image/cached_network_image.dart';
-import 'package:flutter/cupertino.dart';
+import 'package:fluentui_system_icons/fluentui_system_icons.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:go_router/go_router.dart';
-import 'package:gyawun/core/network/internet_guard.dart';
+import 'package:gyawun/core/extensions/random_material_shape.dart';
+import 'package:gyawun/core/widgets/internet_guard.dart';
 import 'package:gyawun/core/utils/service_locator.dart';
-import 'package:gyawun/utils/extensions.dart';
-import 'package:gyawun/utils/playlist_thumbnail.dart';
-
+import 'package:gyawun/core/widgets/library_tile.dart';
+import 'package:gyawun/core/widgets/rounded_polygon_icon.dart';
+import 'package:gyawun/screens/settings/widgets/color_icon.dart';
+import 'package:gyawun/themes/text_styles.dart';
 import '../../../../generated/l10n.dart';
 import '../../../../services/library.dart';
-import '../../../../themes/colors.dart';
 import '../../../../utils/adaptive_widgets/adaptive_widgets.dart';
 import '../../../../utils/bottom_modals.dart';
 import 'cubit/library_cubit.dart';
@@ -28,28 +29,24 @@ class LibraryPage extends StatelessWidget {
         builder: (context, state) {
           return InternetGuard(
             child: Scaffold(
-              appBar: AppBar(
-                title: Text(S.of(context).Saved),
-                centerTitle: true,
-                automaticallyImplyLeading: false,
-                actions: [
-                  IconButton(
+              floatingActionButton: Column(
+                mainAxisSize: .min,
+                children: [
+                  FloatingActionButton.small(
+                    heroTag: 'import_playlist',
                     onPressed: () {
                       Modals.showImportplaylistModal(context);
                     },
-                    icon: const Icon(
-                      Icons.import_export_outlined,
-                      size: 25,
-                    ),
+                    child: const Icon(Icons.import_export_rounded),
                   ),
-                  IconButton(
+                  SizedBox(height: 8),
+                  FloatingActionButton(
+                    heroTag: 'create_playlist',
+
                     onPressed: () {
                       Modals.showCreateplaylistModal(context);
                     },
-                    icon: const Icon(
-                      Icons.add,
-                      size: 25,
-                    ),
+                    child: const Icon(FluentIcons.add_24_filled),
                   ),
                 ],
               ),
@@ -60,7 +57,7 @@ class LibraryPage extends StatelessWidget {
                   :final playlists,
                   :final favouritesCount,
                   :final downloadsCount,
-                  :final historyCount
+                  :final historyCount,
                 ) =>
                   _LibraryBody(
                     playlists: playlists,
@@ -78,11 +75,12 @@ class LibraryPage extends StatelessWidget {
 }
 
 class _LibraryBody extends StatelessWidget {
-  const _LibraryBody(
-      {required this.playlists,
-      this.favouritesCount = 0,
-      this.downloadsCount = 0,
-      this.historyCount = 0});
+  const _LibraryBody({
+    required this.playlists,
+    this.favouritesCount = 0,
+    this.downloadsCount = 0,
+    this.historyCount = 0,
+  });
 
   final Map playlists;
   final int favouritesCount;
@@ -91,169 +89,231 @@ class _LibraryBody extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return SingleChildScrollView(
-      child: Center(
-        child: Container(
-          constraints: const BoxConstraints(maxWidth: 1000),
-          padding: const EdgeInsets.symmetric(horizontal: 8),
-          child: Column(
-            children: [
-              /// FAVOURITES
-              AdaptiveListTile(
-                margin: const EdgeInsets.symmetric(vertical: 4),
-                title: Text(S.of(context).Favourites),
-                leading: _iconBox(context, AdaptiveIcons.heart_fill),
-                subtitle: Text(S.of(context).nSongs(favouritesCount)),
-                trailing: Icon(AdaptiveIcons.chevron_right),
-                onTap: () => context.push('/saved/favourites_page'),
+    return NestedScrollView(
+      headerSliverBuilder: (context, innerBoxIsScrolled) {
+        return [
+          SliverAppBar(
+            pinned: true,
+            expandedHeight: 120,
+            flexibleSpace: FlexibleSpaceBar(
+              titlePadding: .only(left: 16, bottom: 12),
+              title: Column(
+                mainAxisSize: MainAxisSize.min,
+                crossAxisAlignment: CrossAxisAlignment.start,
+                children: [
+                  Text(
+                    'Library',
+                    maxLines: 1,
+                    overflow: TextOverflow.ellipsis,
+                    style: textStyle(context).copyWith(fontSize: 24),
+                  ),
+                ],
               ),
-
-              /// DOWNLOADS
-              AdaptiveListTile(
-                margin: const EdgeInsets.symmetric(vertical: 4),
-                title: Text(S.of(context).Downloads),
-                leading: _iconBox(context, AdaptiveIcons.download),
-                subtitle: Text(S.of(context).nSongs(downloadsCount)),
-                trailing: Icon(AdaptiveIcons.chevron_right),
-                onTap: () => context.push('/saved/downloads_page'),
-              ),
-
-              /// HISTORY
-              AdaptiveListTile(
-                margin: const EdgeInsets.symmetric(vertical: 4),
-                title: Text(S.of(context).History),
-                leading: _iconBox(context, Icons.history),
-                subtitle: Text(S.of(context).nSongs(historyCount)),
-                trailing: Icon(AdaptiveIcons.chevron_right),
-                onTap: () => context.push('/saved/history_page'),
-              ),
-
-              /// PLAYLISTS
-              Column(
-                children: SplayTreeMap.from(playlists)
-                    .map((key, item) {
-                      if (item == null) {
-                        return MapEntry(key, const SizedBox());
-                      }
-
-                      return MapEntry(
-                        key,
-                        AdaptiveListTile(
-                          margin: const EdgeInsets.symmetric(vertical: 4),
-                          title: Text(
-                            item['title'],
-                            maxLines: 2,
-                          ),
-                          leading: _playlistLeading(
-                            context,
-                            key,
-                            item,
-                          ),
-                          subtitle:
-                              (item['songs'] != null || item['isPredefined'])
-                                  ? Text(
-                                      item['isPredefined'] == true
-                                          ? item['subtitle']
-                                          : S.of(context).nSongs(
-                                                item['songs'].length,
-                                              ),
-                                      maxLines: 1,
-                                    )
-                                  : null,
-                          trailing: Icon(AdaptiveIcons.chevron_right),
-                          onTap: () {
-                            if (item['isPredefined'] == true) {
-                              context.push(
-                                '/browse',
-                                extra: {
-                                  'endpoint':
-                                      item['endpoint'].cast<String, dynamic>(),
-                                },
-                              );
-                            } else {
-                              context.push(
-                                '/saved/playlist_details',
-                                extra: {
-                                  'playlistkey': key,
-                                },
-                              );
-                            }
-                          },
-                          onSecondaryTap: () =>
-                              _showPlaylistMenu(context, key, item),
-                          onLongPress: () =>
-                              _showPlaylistMenu(context, key, item),
-                        ),
-                      );
-                    })
-                    .values
-                    .toList(),
-              ),
-            ],
+            ),
           ),
-        ),
+        ];
+      },
+      body: CustomScrollView(
+        slivers: [
+          SliverPadding(
+            padding: const EdgeInsets.symmetric(horizontal: 8.0),
+            sliver: SliverToBoxAdapter(
+              child: Padding(
+                padding: const .symmetric(vertical: 4),
+                child: Column(
+                  children: [
+                    // FAVOURITES
+                    Padding(
+                      padding: const EdgeInsets.symmetric(
+                        horizontal: 8,
+                        vertical: 4,
+                      ),
+                      child: LibraryTile(
+                        title: Text(
+                          S.of(context).Favourites,
+                          style: textStyle(context).copyWith(fontSize: 16),
+                          overflow: .ellipsis,
+                           maxLines: 1,
+                        ),
+                        leading: ColorIcon(
+                          icon: FluentIcons.heart_24_filled,
+                          color: Theme.of(context).colorScheme.primaryContainer,
+                          size: 30,
+                        ),
+                        subtitle: Text(
+                          S.of(context).nSongs(favouritesCount),
+                          style: textStyle(context).copyWith(fontSize: 11),
+                          overflow: .ellipsis,
+                           maxLines: 1,
+                        ),
+                        trailing: Icon(FluentIcons.chevron_right_24_filled),
+                        onTap: () => context.push('/library/favourites'),
+                      ),
+                    ),
+                    // DOWNLOADS
+                    Padding(
+                      padding: const EdgeInsets.symmetric(
+                        horizontal: 8,
+                        vertical: 4,
+                      ),
+                      child: LibraryTile(
+                        title: Text(
+                          S.of(context).Downloads,
+                          style: textStyle(context).copyWith(fontSize: 16),
+                          overflow: .ellipsis,
+                           maxLines: 1,
+                        ),
+                        leading: ColorIcon(
+                          icon: FluentIcons.cloud_arrow_down_24_filled,
+                          color: Theme.of(context).colorScheme.primaryContainer,
+                          size: 30,
+                        ),
+                        subtitle: Text(
+                          S.of(context).nSongs(downloadsCount),
+                          style: textStyle(context).copyWith(fontSize: 11),
+                          overflow: .ellipsis,
+                           maxLines: 1,
+                        ),
+                        trailing: Icon(FluentIcons.chevron_right_24_filled),
+                        onTap: () => context.push('/library/downloads'),
+                      ),
+                    ),
+
+                    /// HISTORY
+                    Padding(
+                      padding: const EdgeInsets.symmetric(
+                        horizontal: 8,
+                        vertical: 4,
+                      ),
+                      child: LibraryTile(
+                        title: Text(
+                          S.of(context).History,
+                          style: textStyle(context).copyWith(fontSize: 16),
+                          overflow: .ellipsis,
+                           maxLines: 1,
+                        ),
+                        leading: ColorIcon(
+                          icon: FluentIcons.history_24_filled,
+                          color: Theme.of(context).colorScheme.primaryContainer,
+                          size: 30,
+                        ),
+                        subtitle: Text(
+                          S.of(context).nSongs(historyCount),
+                          style: textStyle(context).copyWith(fontSize: 11),
+                           overflow: .ellipsis,
+                           maxLines: 1,
+                        ),
+                        trailing: Icon(FluentIcons.chevron_right_24_filled),
+                        onTap: () => context.push('/library/history'),
+                      ),
+                    ),
+                    Column(
+                      children: SplayTreeMap.from(playlists)
+                          .map((key, item) {
+                            if (item == null) {
+                              return MapEntry(key, const SizedBox());
+                            }
+                            return MapEntry(
+                              key,
+                              Padding(
+                                padding: const EdgeInsets.symmetric(
+                                  horizontal: 8,
+                                  vertical: 4,
+                                ),
+                                child: LibraryTile(
+                                  title: Text(
+                                    item['title'],
+                                    maxLines: 1,
+                                    overflow: .ellipsis,
+                                    style: textStyle(
+                                      context,
+                                    ).copyWith(fontSize: 16),
+                                  ),
+                                  leading: _playlistLeading(context, key, item),
+                                  subtitle:
+                                      (item['songs'] != null ||
+                                          item['isPredefined'])
+                                      ? Text(
+                                          item['isPredefined'] == true
+                                              ? item['subtitle']
+                                              : S
+                                                    .of(context)
+                                                    .nSongs(
+                                                      item['songs'].length,
+                                                    ),
+                                          maxLines: 1,
+                                          style: textStyle(
+                                            context,
+                                          ).copyWith(fontSize: 11),
+                                        )
+                                      : null,
+                                  trailing: Icon(
+                                    FluentIcons.chevron_right_24_filled,
+                                  ),
+                                  onTap: () {
+                                    if (item['isPredefined'] == true) {
+                                      context.push(
+                                        '/browse',
+                                        extra: {
+                                          'endpoint': item['endpoint']
+                                              .cast<String, dynamic>(),
+                                        },
+                                      );
+                                    } else {
+                                      context.push(
+                                        '/library/playlist_details',
+                                        extra: {'playlistkey': key},
+                                      );
+                                    }
+                                  },
+                                  onLongPress: () =>
+                                      _showPlaylistMenu(context, key, item),
+                                ),
+                              ),
+                            );
+                          })
+                          .values
+                          .toList(),
+                    ),
+                  ],
+                ),
+              ),
+            ),
+          ),
+        ],
       ),
     );
   }
 
-  Widget _iconBox(BuildContext context, IconData icon) {
-    return Container(
-      height: 50,
-      width: 50,
-      decoration: BoxDecoration(
-        color: greyColor,
-        borderRadius: BorderRadius.circular(3),
-      ),
-      child: Icon(
-        icon,
-        color: context.isDarkMode ? Colors.white : Colors.black,
-      ),
-    );
-  }
-
-  Widget _playlistLeading(
-    BuildContext context,
-    String key,
-    Map item,
-  ) {
+  Widget _playlistLeading(BuildContext context, String key, Map item) {
     if (item['isPredefined'] == true) {
       return ClipRRect(
-        borderRadius: BorderRadius.circular(
-          item['type'] == 'ARTIST' ? 50 : 3,
-        ),
+        borderRadius: BorderRadius.circular(item['type'] == 'ARTIST' ? 30 : 8),
         child: CachedNetworkImage(
-          imageUrl: item['thumbnails']
-              .first['url']
-              .replaceAll('w540-h225', 'w60-h60'),
-          height: 50,
-          width: 50,
+          imageUrl: item['thumbnails'].first['url'].replaceAll(
+            'w540-h225',
+            'w60-h60',
+          ),
+          height: 40,
+          width: 40,
         ),
       );
     }
-
-    if (item['songs'] != null && item['songs'].isNotEmpty) {
-      return PlaylistThumbnail(
-        playslist: item['songs'],
-        size: 50,
-        radius: item['type'] == 'ARTIST' ? 50 : 8,
-      );
-    }
-
-    return _iconBox(context, CupertinoIcons.music_note_list);
+    return Container(
+      padding: const EdgeInsets.all(6),
+      decoration: BoxDecoration(
+        color: Theme.of(context).colorScheme.primaryContainer,
+        borderRadius: BorderRadius.circular(8),
+      ),
+      child: RoundedPolygonIcon(polygon: RandomMaterialShape.random, size: 30),
+    );
   }
 
-  void _showPlaylistMenu(
-    BuildContext context,
-    String key,
-    Map item,
-  ) {
+  void _showPlaylistMenu(BuildContext context, String key, Map item) {
     if (item['videoId'] == null && item['playlistId'] != null) {
       Modals.showPlaylistBottomModal(context, item);
     } else if (item['isPredefined'] == false) {
-      Modals.showPlaylistBottomModal(
-        context,
-        {...item, 'playlistId': key},
-      );
+      Modals.showPlaylistBottomModal(context, {...item, 'playlistId': key});
     }
   }
 }
