@@ -1,5 +1,6 @@
 import 'dart:io';
 
+import 'package:fluentui_system_icons/fluentui_system_icons.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
@@ -23,13 +24,6 @@ class SettingsPage extends StatelessWidget {
     return BlocProvider(
       create: (_) => SettingsSystemCubit()..load(),
       child: AdaptiveScaffold(
-        appBar: AdaptiveAppBar(
-          title: Text(
-            S.of(context).Settings,
-            style: mediumTextStyle(context, bold: false),
-          ),
-          automaticallyImplyLeading: false,
-        ),
         body: BlocBuilder<SettingsSystemCubit, SettingsSystemState>(
           builder: (context, state) {
             final bool? batteryDisabled = state is SettingsSystemLoaded
@@ -39,74 +33,94 @@ class SettingsPage extends StatelessWidget {
             return Center(
               child: Container(
                 constraints: const BoxConstraints(maxWidth: 1000),
-                child: ListView(
-                  padding: const EdgeInsets.symmetric(
-                    horizontal: 16,
-                    vertical: 8,
+                child: NestedScrollView(
+                  headerSliverBuilder: (context, innerBoxIsScrolled) {
+                    return [
+                      SliverAppBar(
+                        pinned: true,
+                        expandedHeight: 120,
+                        flexibleSpace: FlexibleSpaceBar(
+                          titlePadding: .only(left: 16, bottom: 12),
+                          title: Column(
+                            mainAxisSize: MainAxisSize.min,
+                            crossAxisAlignment: CrossAxisAlignment.start,
+                            children: [
+                              Text(
+                                S.of(context).Settings,
+                                maxLines: 1,
+                                overflow: TextOverflow.ellipsis,
+                                style: textStyle(
+                                  context,
+                                ).copyWith(fontSize: 24),
+                              ),
+                            ],
+                          ),
+                        ),
+                      ),
+                    ];
+                  },
+                  body: ListView(
+                    padding: .only(left: 16,right: 16,bottom: 16),
+                    children: [
+                      if (Platform.isAndroid && batteryDisabled != true)
+                        _BatteryWarningTile(),
+                      GroupTitle(title: "General"),
+                      SettingTile(
+                        title: S.of(context).Appearence,
+                        leading: const Icon(FluentIcons.color_background_24_filled),
+                        isFirst: true,
+                        onTap: () => context.go('/settings/appearance'),
+                      ),
+                      SettingTile(
+                        title: "Player",
+                        leading: const Icon(FluentIcons.play_24_filled),
+                        isLast: true,
+                        onTap: () => context.go('/settings/player'),
+                      ),
+                      GroupTitle(title: "Services"),
+                      SettingTile(
+                        title: "Youtube Music",
+                        leading: const Icon(Icons.play_circle_fill),
+                        isFirst: true,
+                        isLast: true,
+                        onTap: () => context.go('/settings/services/ytmusic'),
+                      ),
+                      GroupTitle(title: "Storage & Privacy"),
+                      SettingTile(
+                        title: "Backup and storage",
+                        leading: const Icon(FluentIcons.storage_24_filled),
+                        isFirst: true,
+                        onTap: () => context.go('/settings/backup_storage'),
+                      ),
+                      SettingTile(
+                        title: "Privacy",
+                        leading: const Icon(FluentIcons.shield_keyhole_24_filled),
+                        isLast: true,
+                        onTap: () => context.go('/settings/privacy'),
+                      ),
+                      GroupTitle(title: "Updates & About"),
+                      SettingTile(
+                        title: S.of(context).About,
+                        leading: const Icon(FluentIcons.info_24_filled),
+                        isFirst: true,
+                        onTap: () => context.go('/settings/about'),
+                      ),
+                      SettingTile(
+                        title: S.of(context).Check_For_Update,
+                        leading: const Icon(FluentIcons.arrow_circle_up_24_filled),
+                        onTap: () async {
+                          await UpdateService.manualCheck(context);
+                        },
+                      ),
+                      SettingTile(
+                        leading: const Icon(FluentIcons.money_24_filled),
+                        title: S.of(context).Donate,
+                        isLast: true,
+                        subtitle: S.of(context).Donate_Message,
+                        onTap: () => showPaymentsModal(context),
+                      ),
+                    ],
                   ),
-                  children: [
-                    if (Platform.isAndroid && batteryDisabled != true)
-                      _BatteryWarningTile(),
-                    GroupTitle(title: "General"),
-                    SettingTile(
-                      title: S.of(context).Appearence,
-                      leading: const Icon(Icons.looks_rounded),
-                      isFirst: true,
-                      onTap: () => context.go('/settings/appearance'),
-                    ),
-                    SettingTile(
-                      title: "Player",
-                      leading: const Icon(Icons.play_arrow_rounded),
-                      isLast: true,
-                      onTap: () => context.go('/settings/player'),
-                    ),
-                    GroupTitle(title: "Services"),
-                    SettingTile(
-                      title: "Youtube Music",
-                      leading: const Icon(Icons.play_circle_fill),
-                      isFirst: true,
-                      isLast: true,
-                      onTap: () => context.go('/settings/services/ytmusic'),
-                    ),
-                    GroupTitle(title: "Storage & Privacy"),
-                    SettingTile(
-                      title: "Backup and storage",
-                      leading: const Icon(
-                        Icons.cloud_upload_rounded,
-                      ),
-                      isFirst: true,
-                      onTap: () => context.go(
-                        '/settings/backup_storage',
-                      ),
-                    ),
-                    SettingTile(
-                      title: "Privacy",
-                      leading: const Icon(Icons.privacy_tip),
-                      isLast: true,
-                      onTap: () => context.go('/settings/privacy'),
-                    ),
-                    GroupTitle(title: "Updates & About"),
-                    SettingTile(
-                      title: S.of(context).About,
-                      leading: const Icon(Icons.info_rounded),
-                      isFirst: true,
-                      onTap: () => context.go('/settings/about'),
-                    ),
-                    SettingTile(
-                      title: S.of(context).Check_For_Update,
-                      leading: const Icon(Icons.update_rounded),
-                      onTap: () async {
-                        await UpdateService.manualCheck(context);
-                      },
-                    ),
-                    SettingTile(
-                      leading: const Icon(Icons.money),
-                      title: S.of(context).Donate,
-                      isLast: true,
-                      subtitle: S.of(context).Donate_Message,
-                      onTap: () => showPaymentsModal(context),
-                    ),
-                  ],
                 ),
               ),
             );
@@ -122,26 +136,18 @@ class _BatteryWarningTile extends StatelessWidget {
   Widget build(BuildContext context) {
     return ListTile(
       tileColor: Theme.of(context).colorScheme.errorContainer.withAlpha(200),
-      shape: RoundedRectangleBorder(
-        borderRadius: BorderRadius.circular(25),
-      ),
-      leading: const ColorIcon(
-        icon: Icons.battery_alert,
-        color: Colors.red,
-      ),
+      shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(25)),
+      leading: const ColorIcon(icon: Icons.battery_alert, color: Colors.red),
       title: Text(
         S.of(context).Battery_Optimisation_title,
-        style: TextStyle(
-          color: Theme.of(context).colorScheme.onErrorContainer,
-        ),
+        style: TextStyle(color: Theme.of(context).colorScheme.onErrorContainer),
       ),
       subtitle: Text(
         S.of(context).Battery_Optimisation_message,
         style: tinyTextStyle(context).copyWith(
-          color: Theme.of(context)
-              .colorScheme
-              .onErrorContainer
-              .withValues(alpha: 0.7),
+          color: Theme.of(
+            context,
+          ).colorScheme.onErrorContainer.withValues(alpha: 0.7),
         ),
       ),
       onTap: () {
@@ -154,10 +160,7 @@ class _BatteryWarningTile extends StatelessWidget {
 void showPaymentsModal(BuildContext context) {
   Widget title = AdaptiveListTile(
     contentPadding: EdgeInsets.zero,
-    title: Text(
-      S.of(context).Payment_Methods,
-      style: mediumTextStyle(context),
-    ),
+    title: Text(S.of(context).Payment_Methods, style: mediumTextStyle(context)),
     leading: Column(
       mainAxisAlignment: MainAxisAlignment.center,
       children: [
@@ -174,8 +177,9 @@ void showPaymentsModal(BuildContext context) {
     children: [
       AdaptiveListTile(
         leading: ClipRRect(
-            borderRadius: BorderRadius.circular(6),
-            child: Image.asset('assets/images/upi.jpg', height: 30, width: 30)),
+          borderRadius: BorderRadius.circular(6),
+          child: Image.asset('assets/images/upi.jpg', height: 30, width: 30),
+        ),
         title: Text(
           S.of(context).Pay_With_UPI,
           style: subtitleTextStyle(context),
@@ -187,23 +191,19 @@ void showPaymentsModal(BuildContext context) {
           );
           if (context.mounted) {
             ScaffoldMessenger.of(context).showSnackBar(
-              SnackBar(
-                content: Text(
-                  "Copied UPI ID to clipboard!",
-                ),
-              ),
+              SnackBar(content: Text("Copied UPI ID to clipboard!")),
             );
           }
         },
       ),
       AdaptiveListTile(
         leading: Container(
-            decoration: BoxDecoration(
-              color: const Color.fromRGBO(19, 195, 255, 1),
-              borderRadius: BorderRadius.circular(6),
-            ),
-            child:
-                Image.asset('assets/images/kofi.png', height: 30, width: 30)),
+          decoration: BoxDecoration(
+            color: const Color.fromRGBO(19, 195, 255, 1),
+            borderRadius: BorderRadius.circular(6),
+          ),
+          child: Image.asset('assets/images/kofi.png', height: 30, width: 30),
+        ),
         title: Text(
           S.of(context).Support_Me_On_Kofi,
           style: subtitleTextStyle(context),
@@ -218,9 +218,9 @@ void showPaymentsModal(BuildContext context) {
       ),
       AdaptiveListTile(
         leading: ClipRRect(
-            borderRadius: BorderRadius.circular(6),
-            child:
-                Image.asset('assets/images/coffee.png', height: 30, width: 30)),
+          borderRadius: BorderRadius.circular(6),
+          child: Image.asset('assets/images/coffee.png', height: 30, width: 30),
+        ),
         title: Text(
           S.of(context).Buy_Me_A_Coffee,
           style: subtitleTextStyle(context),
