@@ -5,10 +5,12 @@ class ColorIcon extends StatelessWidget {
   const ColorIcon({
     required this.icon,
     required this.color,
+    this.size,
     super.key,
   });
   final IconData icon;
   final Color? color;
+  final double? size;
 
   @override
   Widget build(BuildContext context) {
@@ -21,7 +23,7 @@ class ColorIcon extends StatelessWidget {
           color: (color != null || context.isDarkMode)
               ? Colors.white.withAlpha(color != null ? 255 : 200)
               : Colors.black.withAlpha(200),
-          size: 20,
+          size:size?? 20,
         ));
   }
 }
