@@ -51,7 +51,7 @@ class MediaPlayer extends ChangeNotifier {
     GetIt.I.registerSingleton<AndroidLoudnessEnhancer>(_loudnessEnhancer);
     if (Platform.isAndroid && _equalizer != null) {
       GetIt.I.registerSingleton<AndroidEqualizer>(_equalizer!);
-      print(GetIt.I<AndroidEqualizer>());
+
     }
 
     _init();
