@@ -1,7 +1,7 @@
 import 'dart:convert';
 import 'dart:developer';
 
-pprint(data) {
+void pprint(Object? data) {
   const JsonEncoder encoder = JsonEncoder.withIndent('  ');
   final jsonString = encoder.convert(data);
   log(jsonString);
